% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_bnf_class.R
\name{count_bnf_class}
\alias{count_bnf_class}
\title{Construct a df summarising number of prescriptions at specific level of BNF}
\usage{
count_bnf_class(df, start_date, end_date, bnf_level = 2)
}
\arguments{
\item{df}{dataframe of pis data in common format}

\item{start_date}{start date for period of interest}

\item{end_date}{end date for period of interest}

\item{bnf_level}{Level of bnf code to summarise at (e.g. 2 = '01', 4 = '0101', 15='0101010G0BNAABY', etc)}
}
\value{
data_frame (prochino, n1, n2, n3, nY) n1 - nY = number of distinct *days* on which a prescription within a specific level/section of the BNF has been prescribed.
}
\description{
Construct a df summarising number of prescriptions at specific level of BNF
}
