% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ever_used.R
\name{ever_used}
\alias{ever_used}
\title{Report whether a drug appears between start and end date}
\usage{
ever_used(df, bnf_code, start_date, end_date)
}
\arguments{
\item{df}{dataframe of pis data in common format}

\item{bnf_code}{bnf code for drug to the level of interest (e.g. '01', '0101', '0101010G0BNAABY', etc)}

\item{end_date}{end date for period of interest}
}
\value{
data_frame (prochi, ???)
}
\description{
Report whether a drug appears between start and end date
}
