% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polypharmacy.R
\name{polypharmacy}
\alias{polypharmacy}
\title{Construct an index of polypharmacy.}
\usage{
polypharmacy(df, start_date, end_date, bnf_level = 6)
}
\arguments{
\item{df}{dataframe of pis data in common format}

\item{start_date}{start date for period of interest}

\item{end_date}{end date for period of interest}

\item{bnf_level}{Level of bnf code to index at. Default is across chapters.}
}
\value{
data frame (prochino, poly_index)
}
\description{
Here the default is assumed to be number of bnf chapters represented in timeframe. poly_index will reflect the total number of sections represented at the requested level. For chapter level, this can be 0-23. Other levels will have different possible totals.
}
